--- resell-shop-api/src/main/kotlin/com/hyuuny/resellshop/bids/service/BidHistoryWriterImpl.kt
@@ -1,24 +1,0 @@
-package com.hyuuny.resellshop.bids.service
-
-import com.hyuuny.resellshop.bids.dataaccess.BidHistoryRepository
-import com.hyuuny.resellshop.bids.domain.BidHistory
-import org.springframework.stereotype.Component
-
-@Component
-class BidHistoryWriterImpl(
-    private val repository: BidHistoryRepository,
-) : BidHistoryWriter {
-
-    override fun insert(insertBidHistory: InsertBidHistory): BidHistory {
-        return repository.save(
-            BidHistory.of(
-                bidId = insertBidHistory.bidId,
-                type = insertBidHistory.type,
-                status = insertBidHistory.status,
-                userId = insertBidHistory.userId,
-                createdAt = insertBidHistory.createdAt,
-            )
-        )
-    }
-
-}
