--- resell-shop-api/src/main/kotlin/com/hyuuny/resellshop/bids/service/BidReaderImpl.kt
@@ -1,21 +1,0 @@
-package com.hyuuny.resellshop.bids.service
-
-import com.hyuuny.resellshop.bids.dataaccess.BidRepository
-import com.hyuuny.resellshop.bids.domain.Bid
-import com.hyuuny.resellshop.core.common.exception.BidNotFoundException
-import org.springframework.data.repository.findByIdOrNull
-import org.springframework.stereotype.Component
-
-@Component
-class BidReaderImpl(
-    private val repository: BidRepository,
-) : BidReader {
-
-    override fun read(id: Long): Bid =
-        repository.findByIdOrNull(id) ?: throw BidNotFoundException("입찰 내역을 찾을 수 없습니다. id: $id")
-
-    override fun read(ids: Collection<Long>): List<Bid> = repository.findAllById(ids)
-
-    override fun readMinPriceProductSizes(productSizeIds: Collection<Long>): List<BidPriceDetailsResponse> =
-        repository.findAllMinPriceByProductSizeIdIn(productSizeIds)
-}
