+++ resell-shop-api/src/main/kotlin/com/hyuuny/resellshop/categories/service/CategoryReader.kt
@@ -1,11 +1,18 @@
 package com.hyuuny.resellshop.categories.service
 
+import com.hyuuny.resellshop.categories.dataaccess.CategoryRepository
 import com.hyuuny.resellshop.categories.domain.Category
+import com.hyuuny.resellshop.core.common.exception.CategoryNotFoundException
+import org.springframework.data.repository.findByIdOrNull
+import org.springframework.stereotype.Component
 
-interface CategoryReader {
+@Component
+class CategoryReader(
+    private val repository: CategoryRepository,
+) {
 
-    fun read(id: Long): Category
+    fun read(id: Long): Category = repository.findByIdOrNull(id)
+        ?: throw CategoryNotFoundException("카테고리를 찾을 수 없습니다. id: $id")
 
-    fun readAll(): List<Category>
-
+    fun readAll(): List<Category> = repository.findAll()
 }
